Use the same set of exports as regular users in shared tests
The shared tests which access `protobuf_upb` or `protobuf_cpp`
have access to more items than the `protobuf` library itself.
This is because the former don't go through the same re-exporting based
on kernel.

I fix this by creating two test-only libraries that perform the same re-exporting
as the `protobuf` library, but with the kernel explicitly set, and changing the shared
tests to reference that instead of the inner runtime library.

This is needed to reliably test macros, where item paths are relative to the invocation,
not eagerly checked at the macro source.

PiperOrigin-RevId: 623897154